requirements.txt
absl-py==2.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
grpcio==1.76.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.8.0+cu128
torchvision==0.23.0+cu128
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
zipp==3.23.0
